+++ package.json
@@ -22,7 +22,7 @@
     "SSR"
   ],
   "scripts": {
-    "start": "parcel watch ./demo/index.ts -d dist & nodemon app.js",
+    "start": "parcel build ./demo/index.js -d dist & nodemon app.js",
     "start-ts": "parcel ./demo/index-ts.html -d dist --open",
     "start-js": "parcel ./demo/index.html -d dist --open",
     "build": "rm -rf build && parcel build ./src/index.ts --no-source-maps -d build && cp -r src/index.d.ts build",
