+++ src/DI/injected.ts
@@ -9,7 +9,7 @@ import "reflect-metadata";
  */
 export default function Injected(_constructor: Function): void {
   // 通过反射机制，获取参数类型列表
-  const paramsTypes: Array<Function> = Reflect.getMetadata('design:paramtypes', _constructor);
+  const paramsTypes: Function[] = Reflect.getMetadata('design:paramtypes', _constructor);
   if (paramsTypes && paramsTypes.length) {
       paramsTypes.forEach(v => {
           if (v === _constructor) {
