+++ src/Http/index.ts
@@ -1,10 +1,5 @@
 import axios from 'axios';
-import { Injectable } from '../DI';
 
-
-@Injectable({
-  isSingletonMode: false,
-})
 class NVHttp {
   public get<P = any, R = any>(url: string, params?: P): Promise<R> {
     return new Promise((resolve, reject) => {
