+++ src/index.d.ts
@@ -47,12 +47,22 @@ export type TInjectTokenProvider = {
     provide: any;
     useClass?: Function;
     useValue?: any;
-  };
+};
+
+export type TUseClassProvider = {
+    provide: any;
+    useClass: Function;
+};
+  
+export type TuseValueProvider = {
+    provide: any;
+    useValue: any;
+};
 
 export type TNvModuleOptions = {
     imports?: Function[];
     components: Function[];
-    providers?: (Function | TInjectTokenProvider)[];
+    providers?: (Function | TUseClassProvider | TuseValueProvider)[];
     exports?: Function[];
     bootstrap?: Function;
 };
@@ -94,7 +104,7 @@ export interface IComponent<State = any, Props = any, Vm = any> {
 export interface INvModule {
     $imports?: Function[];
     $components?: Function[];
-    $providers?: (Function | TInjectTokenProvider)[];
+    $providers?: (Function | TUseClassProvider | TuseValueProvider)[];
     $exports?: Function[];
     providerList?: Map<Function | string, Function>;
     bootstrap?: Function;
