+++ src/types/index.ts
@@ -2,9 +2,9 @@ export { ICompile } from './compile';
 export { ICompileUtilForRepeat, ICompileUtil } from './compileUtils';
 export { ComponentList, IComponent } from './component';
 export { IMiddleware, EsRouteObject, IInDiv } from './indiv';
-export { INvModule, TInjectTokenProvider } from './nvModule';
+export { INvModule, TInjectTokenProvider, TUseClassProvider, TuseValueProvider } from './nvModule';
 export { IKeyWatcher } from './keyWatcher';
 export { TRouter, IRouter } from './router';
 export { IUtil } from './utils';
-export { IVnode, TAttributes, IPatchList, IParseToVnode, IDiffVnode, IRenderVnode, IVirtualDOM } from './virtualDOM';
+export { IVnode, TAttributes, IPatchList } from './virtualDOM';
 export { TFnWatcher, TFnRender, IWatcher } from './watcher';
